Use resolved localhost instead of hostname
eredis, when given a host address, tries to parse the address first, and
if it unparseable, it then tries to resolve in IPv6 mode first. As this
succeeds, eredis tries to run on IPv6, but the redis container is
listening on IPv4 only. So use the IPv4 address as the safest default.